+++ xhydro/indicators.py
@@ -1,4 +1,4 @@
-"""Module to compute extreme and non-extreme indicators using xclim's build_indicator_module_from_yaml."""
+"""Module to compute indicators using xclim's build_indicator_module_from_yaml."""
 
 # Special imports from xscen
 from xscen import compute_indicators